SWEG_SocialMediaProject/requirements.txt
Flask==2.2.2
Flask-CORS==4.0.0
werkzeug==2.2.3
psycopg2
python-dotenv